+++ HISTORY.md
@@ -4,6 +4,9 @@
 * **Notable Fixes:**
     * fixed display of continuations capturing variadic inputs
 
+### 2026-07-27
+* symbols: refactored drawing a single 8-spike gear symbol
+
 ### 2026-07-24
 * new dev version
 * blocks: fixed display of continuations capturing variadic inputs